requirements.txt
# Deepfake Detection - Requirements
# Install with: pip install -r requirements.txt

# Core ML
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow>=9.0.0
numpy>=1.23.0

# Augmentations
albumentations>=1.3.0

# Metrics
scikit-learn>=1.2.0

# Progress bars
tqdm>=4.65.0

# Logging (optional, for structured logs)
# structlog>=23.1.0

# For DOCX generation (existing research pipeline)
python-docx>=0.8.11
